+++ packages/api/lambdas/sf-scheduler.js
@@ -10,6 +10,17 @@ const { buildQueueMessageFromTemplate } = require('@cumulus/message/Build');
 
 const Logger = require('@cumulus/logger');
 const logger = new Logger({ sender: '@cumulus/api/lambdas/sf-scheduler' });
+const CmrProviderNotConfiguredMessage = 'all collections must configure a cmr_provider for sf to be scheduled';
+class CMRProviderNotConfiguredError extends Error {
+  constructor(error) {
+    super(`${error.message} ${CmrProviderNotConfiguredMessage}`);
+
+    this.name = 'CMRProviderNotConfiguredError';
+    this.code = error.code;
+
+    Error.captureStackTrace(this, CMRProviderNotConfiguredError);
+  }
+}
 
 const getApiProvider = (providerId) => {
   if (isNil(providerId)) return undefined;
@@ -54,7 +65,13 @@ async function handleScheduleEvent(event) {
   };
 
   const eventCustomMeta = get(event, 'meta', {});
-
+  const templateCmr = get(messageTemplate, 'meta.cmr', {});
+  const { cmrProvider } = collection;
+  if (!cmrProvider) {
+    throw new CMRProviderNotConfiguredError({
+      message: `no cmr_provider found for collection ${collection.name}___${collection.version}`,
+    });
+  }
   const message = buildQueueMessageFromTemplate({
     messageTemplate,
     asyncOperationId: get(event, 'asyncOperationId'),
@@ -63,12 +80,15 @@ async function handleScheduleEvent(event) {
       ...eventCustomMeta,
       collection,
       provider,
+      cmr: {
+        ...templateCmr,
+        provider: cmrProvider,
+      },
     },
     payload: get(event, 'payload', {}),
     workflow,
     executionNamePrefix: event.executionNamePrefix,
   });
-
   logger.info(`Sending message ${JSON.stringify(message)} to queue ${queueUrl}`);
 
   return SQS.sendSQSMessage(queueUrl, message);