+++ packages/db/src/migrations/20260205224950_add_collections_cmr_provider.ts
@@ -1,0 +1,24 @@
+import { Knex } from 'knex';
+
+export const up = async (knex: Knex): Promise<void> => {
+  if (!await knex.schema.hasColumn('collections', 'cmr_provider')) {
+    await knex.schema.table('collections', (table) => {
+      table.text('cmr_provider') // .notNullable() optional until all pieces are updated to work with this
+        .comment('CMR Provider for this collection');
+    });
+  }
+  await knex.raw('CREATE INDEX CONCURRENTLY IF NOT EXISTS collection_cmr_provider_index ON collections(cmr_provider)');
+};
+
+export const down = async (knex: Knex): Promise<void> => {
+  if (await knex.schema.hasColumn('collections', 'cmr_provider')) {
+    await knex.schema.table('collections', (table) => {
+      table.dropColumn('cmr_provider');
+    });
+  }
+  await knex.raw('DROP INDEX IF EXISTS collection_cmr_provider_index');
+};
+
+exports.config = {
+  transaction: false,
+};