+++ packages/types/api/collections.d.ts
@@ -15,21 +15,22 @@ export interface CollectionFile {
 }
 
 export interface PartialCollectionRecord {
+  cmrProvider: string, //optional until all pieces are updated to work with this
+  createdAt?: number,
   duplicateHandling?: DuplicateHandling,
   files?: CollectionFile[],
   granuleId?: string,
   granuleIdExtraction?: string,
   ignoreFilesConfigForDiscovery?: boolean,
+  meta?: object,
   name?: string,
   process?: string,
   reportToEms?: boolean,
   sampleFileName?: string,
   tags?: string[],
-  meta?: object,
+  updatedAt?: number,
   url_path?: string,
   version?: string,
-  createdAt?: number,
-  updatedAt?: number
 }
 
 export interface NewCollectionRecord extends PartialCollectionRecord {